Refactor to move hashing classes/functions into own file
