+++ phyclone/utils/__init__.py
@@ -1,1 +1,2 @@
-from .utils import get_iterator_length, Timer, list_of_np_cache, two_np_arr_cache
+from .utils import get_iterator_length, Timer
+from .hashing_utils import list_of_np_cache, two_np_arr_cache
